Finat till DataAccess

+++ MongoDBLabb/ItemController.cs
@@ -44,29 +44,46 @@ namespace MongoDBLabb
                     {
                         case 1:
                             io.PrintString("Mata in artikelnummer, namn, enhet, antal och märke till produkten.");
-                            ItemModel item = new ItemModel{Id = int.Parse(io.ReadString()), Name = io.ReadString(), Unit = io.ReadString(), Qty = int.Parse(io.ReadString()), Brand = io.ReadString() };
+                            ItemModel item = new ItemModel { Id = int.Parse(io.ReadString()), Name = io.ReadString(), Unit = io.ReadString(), Qty = int.Parse(io.ReadString()), Brand = io.ReadString() };
                             itemDao.Create(item);
                             io.PrintString($"Artikel {item.Id} tillagt");
                             break;
                         case 2:
                             io.PrintString("Mata in vilken artikel du vill uppdatera.");
                             int i = int.Parse(io.ReadString());
-                            io.PrintString("Mata in antalet du vill ändra till.");
-                            int newid = int.Parse(io.ReadString());
-                            itemDao.Update(i,newid);
+                            if (itemDao.ReadOne(i) == null)
+                            {
+                                io.PrintString("Artikeln du vill uppdatera finns inte.");
+                            }
+                            else
+                            {
+                                io.PrintString("Mata in vad du vill ändra till: ");
+                                int newid = int.Parse(io.ReadString());
+                                itemDao.Update(i, newid);
+                                io.PrintString($"Artikel {i} har nu uppdaterat qty till {newid}");
+                            }
                             break;
                         case 3:
-                            itemDao.ReadAll().ForEach(i => {io.PrintInt(i.Id); io.PrintString(i.Name); io.PrintString(i.Unit); io.PrintInt(i.Qty); io.PrintString(i.Brand); io.PrintString(""); });
+                            itemDao.ReadAll().ForEach(i => { io.PrintInt(i.Id); io.PrintString(i.Name); io.PrintString(i.Unit); io.PrintInt(i.Qty); io.PrintString(i.Brand); io.PrintString(""); });
                             break;
                         case 4:
                             io.PrintString("Mata in det artikelnr du vill visa.");
-                            int id = int.Parse(io.ReadString());                      
+                            int id = int.Parse(io.ReadString());
                             io.PrintDocument(itemDao.ReadOne(id));
                             break;
                         case 5:
-                            io.PrintString("Mata in vilket artnr du vill ta bort från databasen");
+                            io.PrintString("Mata in vilket artnr du vill ta bort");
                             id = int.Parse(io.ReadString());
-                            itemDao.Delete(id);
+                            
+                            if(itemDao.ReadOne(id) == null)
+                            {
+                                io.PrintString("Inget artikelnummer hittades");                        
+                            }
+                            else
+                            {
+                                io.PrintString($"Artikel {id} borttagen");                                
+                                itemDao.Delete(id);
+                            }
                             break;
                         case 6:
                             io.Exit();
